+++ src/python.c
@@ -111,7 +111,7 @@ mk_wargs(int argc, char **argv)
   int i; char *saved_locale = NULL;
   wchar_t **args = NULL;
 
-  args = (wchar_t **)PyMem_Malloc((argc+1)*sizeof(wchar_t *));
+  args = (wchar_t **)PyMem_Malloc((size_t)(argc+1)*sizeof(wchar_t *));
   if (!args) goto oom;
 
   saved_locale = strdup(setlocale(LC_ALL, NULL));
@@ -149,7 +149,7 @@ cp_wargs(int argc, wchar_t **args)
 {
   int i; wchar_t **args_copy = NULL;
   if (!args) return NULL;
-  args_copy = (wchar_t **)PyMem_Malloc((argc+1)*sizeof(wchar_t *));
+  args_copy = (wchar_t **)PyMem_Malloc((size_t)(argc+1)*sizeof(wchar_t *));
   if (!args_copy) goto oom;
   for (i=0; i<(argc+1); i++) { args_copy[i] = args[i]; }
   return args_copy;
@@ -273,7 +273,7 @@ _Py_char2wchar(const char* arg, size_t *size)
   *out = 0;
 #endif
   if (size != NULL)
-    *size = out - res;
+    *size = (size_t)(out - res);
   return res;
  oom:
   fprintf(stderr, "out of memory\n");
